Add a generic clear text message sink.
For now the generic clear text message sink just prints the decoded
clear text messages in the stdout. In the future and depending the needs
of the project, the clear text messages can be forwarded at other
services on the same host or via the network.

Also the debug message source block now supports injection of artificial
errors, for more robust decoding of the Morse decoder block.