Add FM demodulation flowgraph with RIGCTL support
The FM demodulation flowgraph uses the high performance Xlating FIR
filter for filtering and frequency centering at the same time. Tuning
the SDR in a frequency different than the frequency of interest, we
avoid the DC offset that the of the SDR. As far the audio sampling rate
concerns, various configurations are supported due to the Polyphase
Filter-bank arbitrary resampler.

 The flowgraph supports also doppler compensation based on RIGCTL
commands. Instead of chaning the SDR frequency, we digitally perform the
proper correction to avoid noise and lost samples during the frequency
re-configuration. The doppler correction mechanism gathers the frequency
changes from the RIGCTL daemon, and predicts the frequency changes
between the RIGCTL messages for better and more fine-grained correction.