Improve the generic FM demodulator
The FM demodulator now supports a variaty of hardware setups and saves
the WAV file for further processing. A set of command line arguments can
control most of the flowgraph internals, such as the wav file name and
storage path e.t.c.

TODO: Check the settings for all the available hardware