Fix gnuplot issue
In some cases (crash, inproper termination) a flowgraph failed to
produce a binary file that had exactly the proper number of float
numbers in order the gnuplot script to be able to convert it in a matrix
representation.

The updated script tries to truncate the binary file in such a way so it
can be converted into a matrix. This will allow waterfall images to be
created even after a crash.

System should have the "truncate" command line utility installed which
is part of GNU GNU coreutils.