Change NOAA output naming scheme
NOAA APT decoder will create a file without a timestamp. The
satnogs-client will be resposible to append the proper timestamp.
This change was necessary to avoid the time offset between the start of
the observation timestamp that the APT decoder appends to the filename
and the end of the observation that the client takes into account.